@phadej 2020-10-15
There are three SSE2 prs. Can they be combined into one? I'm confused, and don't have time to cross check how they are different.