@ethercrow 2020-10-15
They optimize three different functions individually. Sure, I can open a combined one.